+++ rust_tests/src/bin/update_min_block_time.rs
@@ -1,0 +1,31 @@
+mod path;
+mod update_transaction_header;
+
+use hex;
+use ledger::{ApduCommand, LedgerApp};
+
+fn main() {
+    let mut key_derivation_path = path::generate_key_derivation_path();
+    let mut update_type = hex::decode("13").unwrap();
+
+    let mut min_block_time = hex::decode("00000F0A00000001").unwrap();
+
+    let mut command_data = Vec::new();
+    command_data.append(&mut key_derivation_path);
+    command_data.append(&mut update_transaction_header::generate_update_transaction_header());
+    command_data.append(&mut update_type);
+    command_data.append(&mut min_block_time);
+
+    let command = ApduCommand {
+        cla: 224,
+        ins: 68,
+        p1: 0,
+        p2: 0,
+        length: command_data.len() as u8,
+        data: command_data
+    };
+
+    let ledger = LedgerApp::new().unwrap();
+    let result = ledger.exchange(command).unwrap();
+    println!("Signature for UpdateMinBlockTime: {}", hex::encode(result.data));
+}