+++ governance-app/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support for update timeout parameters transaction.
+- Support for update minimum block time transaction.
 
 ## 0.0.1